feat: image to base64 when screenshot
