+++ packages/docs/fluent-editor/demos/screenshot/upload.vue
@@ -13,28 +13,6 @@ const TOOLBAR_CONFIG = [
   ['screenshot'],
 ]
 
-// to solve html2canvas get image empty
-const imgToBase64 = (imageUrl: string) => new Promise<string>((resolve, reject) => {
-  let canvas = document.createElement('canvas')
-  let img = new Image()
-  img.crossOrigin = 'Anonymous'
-  img.src = imageUrl
-  img.onload = function () {
-    const ctx = canvas.getContext('2d')
-    if (ctx) {
-      canvas.height = img.height
-      canvas.width = img.width
-      ctx.clearRect(0, 0, canvas.width, canvas.height)
-      ctx.drawImage(img, 0, 0)
-      const dataURL = canvas.toDataURL('image/png', 1)
-      resolve(dataURL)
-    }
-  }
-  img.onerror = function () {
-    reject(new Error('Could not load image at ' + imageUrl))
-  }
-})
-
 onMounted(() => {
   // ssr compat, reference: https://vitepress.dev/guide/ssr-compat#importing-in-mounted-hook
   import('@opentiny/fluent-editor').then((module) => {
@@ -47,13 +25,6 @@ onMounted(() => {
       },
       screenshot: {
         Html2Canvas,
-        onclone: async (doc: Document) => {
-          const imgs = doc.querySelectorAll('img')
-          const promises = Array.from(imgs).map(async (img) => {
-            img.src = await imgToBase64(img.src)
-          })
-          await Promise.all(promises)
-        },
         beforeCreateImage(canvas) {
           return new Promise((resolve) => {
             canvas.toBlob(
